(feat) make it usable on stable
This introduces the `cache` feature which enables the caching
functionality. At the moment this is only usable in nightly rust.

As a side-effect, a request to a file for which the process doesn't have
permission to read yields an HTTP 403: Forbidden status, non-existent
file requests yield an HTTP 404: Not Found status, and any other error
yields an HTTP 500: Internal Server Error status.